Bottom Navigation Views Activity in PokemonDetailsActivity
I changed the kind of activity used in PokemonDetailsActivity, adding 3 fragments, users can navigate between them with the bottom bar.

The first fragment will show basic info (right now it shows everything, I need to move some info to other fragments). The second fragment will show the stats (currently in the first one). And the third fragment will show the moves (yet to do)

Another change is that Pokemon stats are now included in the details, with bars that show the stat value changing the length and color