TeamDetailsActivity, update and remove
When you click on a team, it will open an Activity with its information (for now just the name). There, a toolbar contains menu options to update the name or delete the team, and this will be reflected in both the database and the interface in real time

Also, the option to add a Pokemon to a team now shows the available teams of the user (yet they still do nothing when being added)

I created the class ToastUtil to centralize the way Toasts are made and to avoid repeating code. This also helps to cancel toasts, not filling the screen with them