BasePokemon table + important change regarding how Pokemon data is treated
I optimized the process that I use to get all the Pokemon in the Pokedex. Instead of using the class Pokemon, which stores all the information, I created a BasePokemon class with only the needed data to show them in the list (name, sprite, number, url to get the data later and types for future filters). This is stored in the database, so the app is not constantly making 1025 requests when the user opens the app, and it only does the requests once.

The full data of the Pokemon is now loaded when you click on it, before opening the details, so it only loads the whole information for one Pokemon and not every single one

The recovered Pokemon from the API are stored in blocks of 60, and not when received, to avoid data loss if the user closes the app during the load. This way, if the user stops the app with 520 Pokemon loaded and then opens it again, it will load 520 from the database and then keep recovering the missing ones from the API